Add static `from` methods to configs to take the typesafe configs and convert it to the misq config values.
Use a static `getConfig` method for config factory classes.

Copy Misq config file to test resource folder.

Remove API in NetworkServiceConfigFactory which would not
take a typesafe config.
If a NetworkService.Config should be created without typesafe
config we can create it without the factory class. But we should
avoid that as it would cause duplicate maintenance work when we
change configs.