Replace BalanceListener with observableBalanceAsCoin
Use Coin instead of Double
Use AmountFormatter instead of a custom formatting.

Observable is similar to ObjectProperty but for non-JavaFX usage.
The FxBindings class provides the glue between a JavaFX
property and the Observable from the domain.
Like that:
`balancePin = FxBindings.bind(model.balanceAsCoinProperty).to(walletService.getObservableBalanceAsCoin());`

The value gets applies at binding definition time, so no need to apply the value initially.
The mapping from a non JavaFX thread to the JavaFX thread is done inside FxBindings.

This is a more convenient alternative to using listeners.