Fix exception on closing empty market combobox
Avoid opening an empty combobox when market prices have not yet
populated the combobox list. The empty combobox opens in the top left
corner (on Linux) and throws a list out of bounds exception when closed

I don't have the overview to say if it makes sense to completely prevent
comboboxes from opening without a corresponding list, but there could be
cases where that makes sense.

Cleanup recommendations