Set selectedUserIdentity to an instance from available identities
The expectation when using selectedUserIdentity would be that it's one
of the instances from the set of userIdentities. This would be the case
after selecting a new identity, or creating a new one. Only in the case
of reading from file was the selected identity not an instance from
userIdentities.

I don't think this has been an issue so far, but it seems it could
become one since the expectation is that the instances are the same.