Add handling for events which arrive out of order.
We cannot use the state enum ordinal as we should not rely on the State ordinal. The transition definition is the only relevant data.
Also, State can be implemented in a class as well, it is not mandatory an enum.

The client code is responsible for persisting the 2 hashsets to ensure that pending events are correctly handled after restart.