Unit test payment method accounts
Discovered and fixed some issues along the way. The most important is
the removal of HRK as supported currency in Revolut and Wise. HRK was
replaced by the Euro in 2023. The presence of HRK was preventing the
Revolut and Wise FiatPaymentRails enums from being properly initialized
(they were throwing NullPointerException, I think because HRK wasn't
found in the currency repository).