Refactoring: Rename Listener to Subscriber and rename addListener/removeListener to subscribe/unsubscribe
We will change behaviour of subscribe method to notify subscribers at the subscribe call in the next commit. To reflect that different pattern we renamed it. It follows the behaviour of EasyBind and the Bisq observable implementation (and is in contrast to the JavaFx listener pattern where the client does not get notified at adding a listener).