Add cost factor to messages and DistributedData
We derive the relative cost by TTL and mapSize in relation to their max values the MetaData.
We do not multiply here as that would render most values very low. We use addition and check bounds to stay in the 0-1 bounds.
The applied costs are set according to the expected capability of the sender (e.g. bonded nodes can be expected to be able to perform higher pow).
By that we reduce abuse potential of those class of messages.
Messages/data which have a larger tolerance for its size (e.g. ConfidentialData) gets a higher cost factor.
Also messages which in case of abuse can cause more impact on the receiver like mailbox messages, persisted messages, authenticated messages or messages to bonded roles, get a higher cost factor.

As the calculation of the costs is used for the pow verification we must be careful with changes once deployed.
We use a tolerance in the pow check so slight changes can be tolerated, but its important that we get a good first shot with our educated guesses.