Update filtered markets on new offers
To immediately show markets with new offers that previously had no
offers the filtered markets need to be updated when new offers appear.

The most obvious use case is to allow a user that creates the first
offer for a market to see the market pop up in the filtered list.