Implement Request/Response interfaces for AckRequestingMessage and AckMessage.
There are many classes implementing AckRequestingMessage and therefor inherit the interfaces. We must not confuse it with higher level Request/Response pattern as it is only about responding with the AckMessage. E.g. BisqEasyTakeOfferRequest and BisqEasyTakeOfferResponse would be a higher level Request/Response pattern, but as BisqEasyTakeOfferRequest implements AckRequestingMessage the AckMessage is the actual Response in our lower level framework.

As the Request/Response handling is only used for prioritizing connections at peer management and tracking rrt we do not apply the framework to higher level request/response patterns.