Let Tor pick control port
To select the Tor control port, we bind to a random port, close it and
pass the port number to Tor. It can happen that Tor tries to bind to the
port before it is closed.  In this case, the Tor startup will fail
because Tor will never print a line containing "[notice] Opened Control
listener connection (ready) on ". Up until now, Bisq monitored the Tor
log file to know when the control port is ready. Now Tor writes the
control port to a file and Bisq reads that file.

Ref: #1798