Add @ExcludeForHash to inventory.
When objects get an update we publish both the old version and the new one, where the new one is considered invalid for old nodes. The Invenotry contains all those data and would become itself invalid if we do not exclude the inventory from the hash creation for the pow check.
Note that InventoryResponse is a direct message and its only the pow use case where the hash would cause issues.