Fix bug with cookie subkey handling
We have overwritten the map entry when using different subkeys because enum identity does not consider the values of fields.
We now wrap the key and the subkey in a new object and use that as map key.
We also store the subkey in proto in a new field instead of encoding it in the key.