Add displayName to TradeCurrency and Market.
We have used the name from the fiat currency based on the user's locale which would cause different hashed when used in network messages. To apply ExcludeForHash would require to support versioning in all the objects using those data which would be complicated.
With this change we use the US locale for the name and add a new getDisplayName() method which takes the users locale for the display name. This is a transient field thus it does not cause conflicts.