Add seed nodes after if there are no candidates at extending peer group.
For first start in case all seed nodes are no available, we give up connection to seed nodes after 10 attempts. We try though to extend our peer group but as there are no persisted and reported peers that list is empty. We add now the seednodes in that case to increase changes for recovering. We could increase the retry attempts as well, but it seems that change also improves resilience in case the persisted and reported peers are outdated, so we get fresh nodes from seed nodes.