Add devModeReputationScore to config.
Can be set as jvm argument if isDevMode is true. Value has to be > 0.
All trade apps need to have that value set as it is used in a static manner independent of the user profile.
This makes dev testing easier as sellers need to have reputation to be able to trade.